+++ CHANGELOG.md
@@ -90,6 +90,10 @@
   [Juozas Valancius](https://github.com/juozasvalancius)
   [#3840](https://github.com/realm/SwiftLint/issues/3840)
 
+* Add back `void_function_in_ternary` opt-in rule to warn against using
+  a ternary operator to call `Void` functions.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+
 * Add back `return_value_from_void_function` opt-in rule to warn against using
   `return <expression>` in a function that returns `Void`.  
   [Marcelo Fabri](https://github.com/marcelofabri)